Fixed syntax issue
I am making a GUI ReVanced Patcher which can dynamically include new patches that are stated here, as well as analyse version compatibility by regexing from this document. I would use the ReVanced CLI patch list from terminal, but that doesn't state version compatibility, so I have to stick with this. Please, if it's possible, keep the syntax logic of this readme document.