fix(spoof-wifi-connection): use updated instruction indices
Not using a mutableMethod resulted in using the original method
for locating the instruction indices to patch. When another patch
already patched that method and the indices got offset, we still
used the old indices which resulted in invalid smali.

This is intended to be a temporary fix until a better solution
for mutability is found.

reference: #2177