fix: Use a supplier for literals
This is important, because patches are singletons, thus are directly initialized, causing fingerprints, which are also singletons to be initialized instantly too. If the literal property is consumed instantly, the patches can not execute before, to set the literal properties and the fingerprints will not be able to use them. Use a supplier in order to consume the literal when the fingerprint is resolved.